refactor: update class structures
