+++ paddle/fluid/distributed/collective/process_group_custom.cc
@@ -103,7 +103,9 @@ ProcessGroupCustom::ProcessGroupCustom(const std::shared_ptr<Store>& store,
                                        int rank,
                                        int size,
                                        int gid)
-    : ProcessGroup(rank, size, gid), store_(store), device_type_(device_type) {}
+    : ProcessGroupWithoutStream(rank, size, gid),
+      store_(store),
+      device_type_(device_type) {}
 
 void ProcessGroupCustom::BroadcastUniqueCustomID(
     std::vector<phi::ccl::CCLRootId>& ccl_ids) {  // NOLINT
