+++ paddle/fluid/distributed/collective/process_group_mpi.h
@@ -26,6 +26,7 @@
 #include <mutex>
 
 #include "paddle/fluid/distributed/collective/process_group.h"
+#include "paddle/fluid/distributed/collective/process_group_without_stream.h"
 #include "paddle/fluid/distributed/collective/types.h"
 #include "paddle/fluid/platform/device_context.h"
 
@@ -57,7 +58,7 @@ struct TaskEntry {
   std::function<void(std::unique_ptr<TaskEntry>&)> run_;
 };
 
-class ProcessGroupMPI : public ProcessGroup {
+class ProcessGroupMPI : public ProcessGroupWithoutStream {
  public:
   class MPITask : public ProcessGroup::Task {
    public:
