+++ paddle/fluid/distributed/collective/process_group_nccl.h
@@ -21,6 +21,7 @@
 #include <vector>
 
 #include "paddle/fluid/distributed/collective/process_group.h"
+#include "paddle/fluid/distributed/collective/process_group_with_stream.h"
 #include "paddle/fluid/distributed/store/store.h"
 #include "paddle/fluid/platform/device_event.h"
 #include "paddle/phi/common/place.h"
@@ -39,11 +40,11 @@
 namespace paddle {
 namespace distributed {
 
-using Place = paddle::platform::Place;
+using Place = phi::Place;
 
-class ProcessGroupNCCL final : public ProcessGroup {
+class ProcessGroupNCCL final : public ProcessGroupWithStream {
  public:
-  class NCCLTask final : public ProcessGroup::TaskStream,
+  class NCCLTask final : public ProcessGroupWithStream::TaskStream,
                          public std::enable_shared_from_this<NCCLTask> {
    public:
     NCCLTask(const Place& place,
@@ -94,20 +95,6 @@ class ProcessGroupNCCL final : public ProcessGroup {
       const phi::DenseTensor& in_tensor,
       int64_t offset,
       int64_t numel,
-      bool sync_op) override {
-    return AllGather(out_tensor,
-                     in_tensor,
-                     offset,
-                     numel,
-                     sync_op,
-                     /*use_calc_stream*/ false);
-  }
-
-  std::shared_ptr<ProcessGroup::Task> AllGather(
-      phi::DenseTensor* out_tensor,
-      const phi::DenseTensor& in_tensor,
-      int64_t offset,
-      int64_t numel,
       bool sync_op,
       bool use_calc_stream) override;
 
@@ -115,34 +102,8 @@ class ProcessGroupNCCL final : public ProcessGroup {
       phi::DenseTensor* out_tensor,
       const phi::DenseTensor& in_tensor,
       const AllreduceOptions& opts,
-      bool sync_op) override {
-    return AllReduce(out_tensor,
-                     in_tensor,
-                     opts,
-                     sync_op,
-                     /*use_calc_stream*/ false);
-  }
-
-  std::shared_ptr<ProcessGroup::Task> AllReduce(
-      phi::DenseTensor* out_tensor,
-      const phi::DenseTensor& in_tensor,
-      const AllreduceOptions& opts,
       bool sync_op,
       bool use_calc_stream) override;
-
-  std::shared_ptr<ProcessGroup::Task> AllToAll(
-      phi::DenseTensor* out_tensor,
-      const phi::DenseTensor& in_tensor,
-      const std::vector<int64_t>& out_size_each_rank,
-      const std::vector<int64_t>& in_size_each_rank,
-      bool sync_op) override {
-    return AllToAll(out_tensor,
-                    in_tensor,
-                    out_size_each_rank,
-                    in_size_each_rank,
-                    sync_op,
-                    /*use_calc_stream*/ false);
-  }
 
   std::shared_ptr<ProcessGroup::Task> AllToAll(
       phi::DenseTensor* out_tensor,
@@ -159,31 +120,8 @@ class ProcessGroupNCCL final : public ProcessGroup {
       phi::DenseTensor* out_tensor,
       const phi::DenseTensor& in_tensor,
       const BroadcastOptions& opts,
-      bool sync_op) override {
-    return Broadcast(out_tensor,
-                     in_tensor,
-                     opts,
-                     sync_op,
-                     /*use_calc_stream*/ false);
-  }
-
-  std::shared_ptr<ProcessGroup::Task> Broadcast(
-      phi::DenseTensor* out_tensor,
-      const phi::DenseTensor& in_tensor,
-      const BroadcastOptions& opts,
       bool sync_op,
       bool use_calc_stream) override;
-
-  std::shared_ptr<ProcessGroup::Task> Reduce(phi::DenseTensor* out_tensor,
-                                             const phi::DenseTensor& in_tensor,
-                                             const ReduceOptions& opts,
-                                             bool sync_op) override {
-    return Reduce(out_tensor,
-                  in_tensor,
-                  opts,
-                  sync_op,
-                  /*use_calc_stream*/ false);
-  }
 
   std::shared_ptr<ProcessGroup::Task> Reduce(phi::DenseTensor* out_tensor,
                                              const phi::DenseTensor& in_tensor,
@@ -195,31 +133,8 @@ class ProcessGroupNCCL final : public ProcessGroup {
       phi::DenseTensor* out_tensor,
       const phi::DenseTensor& in_tensor,
       const ReduceScatterOptions& opts,
-      bool sync_op) override {
-    return ReduceScatter(out_tensor,
-                         in_tensor,
-                         opts,
-                         sync_op,
-                         /*use_calc_stream*/ false);
-  }
-
-  std::shared_ptr<ProcessGroup::Task> ReduceScatter(
-      phi::DenseTensor* out_tensor,
-      const phi::DenseTensor& in_tensor,
-      const ReduceScatterOptions& opts,
       bool sync_op,
       bool use_calc_stream) override;
-
-  std::shared_ptr<ProcessGroup::Task> Scatter(phi::DenseTensor* out_tensor,
-                                              const phi::DenseTensor& in_tensor,
-                                              const ScatterOptions& opts,
-                                              bool sync_op) override {
-    return Scatter(out_tensor,
-                   in_tensor,
-                   opts,
-                   sync_op,
-                   /*use_calc_stream*/ false);
-  }
 
   std::shared_ptr<ProcessGroup::Task> Scatter(phi::DenseTensor* out_tensor,
                                               const phi::DenseTensor& in_tensor,
@@ -231,34 +146,8 @@ class ProcessGroupNCCL final : public ProcessGroup {
                                            int src_rank,
                                            int64_t offset,
                                            int64_t numel,
-                                           bool sync_op) override {
-    return Recv(tensor,
-                src_rank,
-                offset,
-                numel,
-                sync_op,
-                /*use_calc_stream*/ false);
-  }
-
-  std::shared_ptr<ProcessGroup::Task> Recv(phi::DenseTensor* tensor,
-                                           int src_rank,
-                                           int64_t offset,
-                                           int64_t numel,
                                            bool sync_op,
                                            bool use_calc_stream) override;
-
-  std::shared_ptr<ProcessGroup::Task> Send(const phi::DenseTensor& tensor,
-                                           int dst_rank,
-                                           int64_t offset,
-                                           int64_t numel,
-                                           bool sync_op) override {
-    return Send(tensor,
-                dst_rank,
-                offset,
-                numel,
-                sync_op,
-                /*use_calc_stream*/ false);
-  }
 
   std::shared_ptr<ProcessGroup::Task> Send(const phi::DenseTensor& tensor,
                                            int dst_rank,
